RF: Move mmap/keep_file_open parameters to DataobjImage.from_file*
